requirements.txt
django==3.0.7
djangorestframework==3.11.0
django-allauth==0.42.0 
django-cors-headers==3.3.0